factor out display settings header
